refactor: separating components logic and libraries into their own file
also, created a new component for reusing the core UI (header, footer, etc)
