+++ src/components/Home.js
@@ -1,0 +1,12 @@
+function Home(props) {
+    return (
+        <>
+            <p>This is the HOME page content</p>
+            <p>Nothing to do here at the moment!</p>
+            <p>Please use the Login feature</p>
+            <p>Have fun!</p>
+        </>
+    );
+}
+
+export default Home;
