+++ src/components/Logout.js
@@ -1,0 +1,9 @@
+'use strict'
+
+function Logout() {
+    return (
+        <p>You have been logged out</p>
+    );
+}
+
+export default Logout;
